All Layouts working
All layouts are working properly.

To implement later:
Long tap gesture for the pictureInPicture option that allows you to change edit modes from zoom/scroll image to moving the image around on the screen. (Version 2.0 maybe includes being able to resize the shape of the picture).
Resize where it splits the image on the other topBottom and leftRight layouts.