Fixed Bug for iphone 4s
Set the photo aspect ratio to 1:1 when it is an iphone 4s (aka when it has a width of 320 and height of 480).

BUG remaining: the launch screen gets cut off on the iphone 4s.
Also, there are three items that need to be fixed in order to be production ready:
- Two lines in the editViewController need to be removed (marked with a "// TODO: remove this line completely")
- The initialViewController is set to the editViewController on the storyboard. This needs to be set back to the cameraController.